+++ package.json
@@ -1,5 +1,11 @@
 {
   "name": "@ekwoka/spotify-api",
+  "type": "module",
+  "files": [
+    "dist"
+  ],
+  "main": "dist/",
+  "types": "dist/",
   "author": "Eric Kwoka",
   "license": "MIT",
   "version": "0.0.1",
@@ -13,31 +19,26 @@
     "typescript",
     "node"
   ],
-  "type": "module",
-  "files": [
-    "dist"
-  ],
   "scripts": {
     "build": "node scripts/esbuild.js",
     "build:tsc": "tsc",
     "lint": "eslint --fix .; prettier --write . --loglevel error",
-    "patch": "pnpm run build:tsc && pnpm version patch && pnpm publish",
-    "minor": "pnpm run build:tsc && pnpm version minor && pnpm publish",
-    "major": "pnpm run build:tsc && pnpm version major && pnpm publish",
+    "patch": "pnpm run test:run && pnpm run build:tsc && pnpm version patch && pnpm publish",
+    "minor": "pnpm run test:run && pnpm run build:tsc && pnpm version minor && pnpm publish",
+    "major": "pnpm run test:run && pnpm run build:tsc && pnpm version major && pnpm publish",
     "test": "vitest",
+    "test:run": "vitest run",
     "coverage": "vitest run --coverage"
-  },
-  "dependencies": {
-    "eslint": "^8.20.0",
-    "vite": "^3.0.2",
-    "vitest": "^0.18.1"
   },
   "devDependencies": {
     "c8": "^7.12.0",
     "esbuild": "^0.14.49",
+    "eslint": "^8.20.0",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.7.1",
-    "typescript": "^4.7.4"
+    "typescript": "^4.7.4",
+    "vite": "^3.0.2",
+    "vitest": "^0.18.1"
   },
   "prettier": {
     "singleQuote": true,
