+++ src/core/index.ts
@@ -1,3 +1,3 @@
-export { client } from "./client";
-export { setToken } from "./setToken";
-export type { Token, QueryFunction, QueryConstructor } from "./client";
+export { setToken } from './setToken';
+export { spotifyApi } from './spotifyApi';
+export type { Token, QueryFunction, QueryConstructor } from './spotifyApi';
