+++ src/core/setToken.ts
@@ -1,4 +1,4 @@
-import { QueryConstructor } from './client';
+import { QueryConstructor } from './spotifyApi';
 
 export const setToken: QueryConstructor<boolean> = (token: string) => (tkn) => {
   tkn.current = token;
