+++ src/index.ts
@@ -1,1 +1,6 @@
-export { SPOTIFY_URL } from "./constants";
+export { SPOTIFY_URL } from './constants';
+export { spotifyApi } from './core/spotifyApi';
+export { getCurrentUser } from './core/endpoints/users/getCurrentUser';
+export { setToken } from './core/setToken';
+export { deepFreeze } from './utils/deepFreeze';
+export type { Token, QueryFunction, QueryConstructor } from './core/spotifyApi';
